Samar

+++ server/database/models/history.js
@@ -9,19 +9,28 @@ function getAll(callback) {
   }
 
 
-  function addHistory (date, type, reciver, card_idcard, users_idusers,amount , callback){
+  function addHistory (date, type, receive, card_idcard,amount , callback){
    
-    conn.query("INSERT INTO history (date, type, reciver, card_idcard, users_idusers,amount ) VALUES (?, ?, ?, ?, ?, ?)" , [date, type, reciver, card_idcard, users_idusers , amount ],function(error , result){
+    conn.query("INSERT INTO history (date, type, receive, card_idcard,amount ) VALUES (?, ?, ?, ?, ?)" , [date, type, receive, card_idcard, amount ],function(error , result){
         callback(error , result)
     })
 
   }
   function deleteAll (callback){
    
-    conn.query("DELETE FROM history ",function(error , result){
+    conn.query("DELETE FROM history ",card_idcard,function(error , result){
         callback(error , result)
     })
 
   }
 
-  module.exports = {getAll , addHistory , deleteAll}
+  function getHistoryByUserId( card_idcard ,callback){
+   
+    conn.query("SELECT *  FROM history WHERE card_idcard= ? ", [card_idcard] , function(error , result){
+        callback(error , result)
+    })
+
+  }
+
+
+  module.exports = {getAll , addHistory , deleteAll , getHistoryByUserId}
